move bcrypt to the into api
