+++ api/cmd/main.go
@@ -5,12 +5,12 @@ import (
 	"github.com/golang-migrate/migrate/v4"
 	"github.com/jackc/pgx/v4"
 	"github.com/lordvidex/gomoney/api/internal/adapters/api"
+	"github.com/lordvidex/gomoney/api/internal/adapters/encryption"
 	mgrpc "github.com/lordvidex/gomoney/api/internal/adapters/grpc"
 	"github.com/lordvidex/gomoney/api/internal/adapters/memory"
 	"github.com/lordvidex/gomoney/api/internal/adapters/paseto"
 	"github.com/lordvidex/gomoney/api/internal/application"
 	"github.com/lordvidex/gomoney/pkg/config"
-	"github.com/lordvidex/gomoney/pkg/encryption"
 	"github.com/pkg/errors"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
